+++ mozglue/baseprofiler/core/ProfileBufferEntry.cpp
@@ -14,7 +14,6 @@
 #include "mozilla/StackWalk.h"
 
 #include "BaseProfiler.h"
-#include "BaseProfilerMarkerPayload.h"
 #include "mozilla/BaseProfilerMarkers.h"
 #include "platform.h"
 #include "ProfileBuffer.h"
@@ -453,9 +452,9 @@ class EntryGetter {
 //     | Label FrameFlags? DynamicStringFragment* LineNumber? CategoryPair?
 //     | JitReturnAddr
 //     )+
-//     Marker*
 //     Responsiveness?
 //   )
+//   | MarkerData
 //   | ( /* Counters */
 //       CounterId
 //       Time
@@ -471,7 +470,7 @@ class EntryGetter {
 //   | Resume
 //   | ( ProfilerOverheadTime /* Sampling start timestamp */
 //       ProfilerOverheadDuration /* Lock acquisition */
-//       ProfilerOverheadDuration /* Expired markers cleaning */
+//       ProfilerOverheadDuration /* Expired data cleaning */
 //       ProfilerOverheadDuration /* Counters */
 //       ProfilerOverheadDuration /* Threads */
 //     )
@@ -594,8 +593,8 @@ void ProfileBuffer::StreamSamplesToJSON(SpliceableJSONWriter& aWriter,
       //
       // - We skip samples that don't have an appropriate ThreadId or Time.
       //
-      // - We skip range Pause, Resume, CollectionStart, Marker, Counter
-      //   and CollectionEnd entries between samples.
+      // - We skip range Pause, Resume, CollectionStart, Counter and
+      //   CollectionEnd entries between samples.
       while (e.Has()) {
         if (e.Get().IsThreadId()) {
           break;
@@ -810,61 +809,22 @@ void ProfileBuffer::StreamMarkersToJSON(SpliceableJSONWriter& aWriter,
     MOZ_ASSERT(static_cast<ProfileBufferEntry::KindUnderlyingType>(type) <
                static_cast<ProfileBufferEntry::KindUnderlyingType>(
                    ProfileBufferEntry::Kind::MODERN_LIMIT));
-    // Code should *break* from the switch if the entry was not fully read.
-    // Code should *return* from the switch if the entry was fully read.
-    switch (type) {
-      case ProfileBufferEntry::Kind::MarkerData:
-        if (aER.ReadObject<int>() != aThreadId) {
-          break;
-        }
-        // Schema:
-        //   [name, time, category, data]
-        aWriter.StartArrayElement();
-        {
-          std::string name = aER.ReadObject<std::string>();
-          const ProfilingCategoryPairInfo& info = GetProfilingCategoryPairInfo(
-              static_cast<ProfilingCategoryPair>(aER.ReadObject<uint32_t>()));
-          auto payload = aER.ReadObject<UniquePtr<ProfilerMarkerPayload>>();
-          double time = aER.ReadObject<double>();
-          MOZ_ASSERT(aER.RemainingBytes() == 0);
-
-          aUniqueStacks.mUniqueStrings->WriteElement(aWriter, name.c_str());
-          aWriter.DoubleElement(time);
-          aWriter.IntElement(unsigned(info.mCategory));
-          if (payload) {
-            aWriter.StartObjectElement(SpliceableJSONWriter::SingleLineStyle);
-            {
-              payload->StreamPayload(aWriter, aProcessStartTime, aUniqueStacks);
-            }
-            aWriter.EndObject();
-          }
-        }
-        aWriter.EndArray();
-        return;
-
-      case ProfileBufferEntry::Kind::Marker:
-        if (::mozilla::base_profiler_markers_detail::
-                DeserializeAfterKindAndStream(
-                    aER, aWriter, aThreadId,
-                    [&](const mozilla::ProfilerString8View& aName) {
-                      aUniqueStacks.mUniqueStrings->WriteElement(
-                          aWriter, aName.String().c_str());
-                    },
-                    [&](ProfileChunkedBuffer& aChunkedBuffer) {
-                      ProfilerBacktrace backtrace("", aThreadId,
-                                                  &aChunkedBuffer);
-                      backtrace.StreamJSON(
-                          aWriter, TimeStamp::ProcessCreation(), aUniqueStacks);
-                    })) {
-          return;
-        }
-        break;
-
-      default:
-        break;
+    if (type != ProfileBufferEntry::Kind::Marker ||
+        !::mozilla::base_profiler_markers_detail::DeserializeAfterKindAndStream(
+            aER, aWriter, aThreadId,
+            [&](const mozilla::ProfilerString8View& aName) {
+              aUniqueStacks.mUniqueStrings->WriteElement(
+                  aWriter, aName.String().c_str());
+            },
+            [&](ProfileChunkedBuffer& aChunkedBuffer) {
+              ProfilerBacktrace backtrace("", aThreadId, &aChunkedBuffer);
+              backtrace.StreamJSON(aWriter, TimeStamp::ProcessCreation(),
+                                   aUniqueStacks);
+            })) {
+      // Not a marker, or marker for another thread.
+      // We probably didn't read the whole entry, so we need to skip to the end.
+      aER.SetRemainingBytes(0);
     }
-
-    aER.SetRemainingBytes(0);
   });
 }
